pyproject.toml
[project]
name = "autonomous-research-agent"
version = "0.1.0"
description = "Production-grade autonomous research agent with hallucination detection, citation tracking, and Discord integration"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # Core LLM & Agent Orchestration
    "langchain>=0.1.0",
    "langgraph>=0.0.1",
    "langchain-groq>=0.1.0",
    
    # Web Scraping & Search
    "beautifulsoup4>=4.12.0",
    "httpx>=0.24.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "lxml>=4.9.0",
    
    # Caching & Database
    "redis>=5.0.0",
    "pymongo>=4.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Discord Integration
    "discord-webhook>=1.3.0",
    "discord.py>=2.4.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic-core>=2.0.0",
    "tenacity>=8.2.0",  # Retry logic
    "typing-extensions>=4.8.0",
    
    # Monitoring & Logging
    "python-json-logger>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pytest-cov>=4.1.0",
    "ipython>=8.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.workspace]
members = [
    "Starting0",
]

[dependency-groups]
dev = [
    "black>=23.7.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "pytest-cov>=4.1.0",
    "ipython>=8.14.0",
]
